fix Instruction entries missing from docset
